requirements.txt
openai>=1.0.0
yfinance>=0.2.0
pandas>=2.0.0
matplotlib>=3.7.0
aiohttp>=3.8.0
html5lib>=1.1
duckduckgo_search>=3.0.0
beautifulsoup4>=4.9.0
requests>=2.31.0
Pillow>=10.0.0
huggingface_hub>=0.19.0
markdownify>=0.11.0
mammoth>=1.6.0
pdfminer.six>=20221105